Externalize the connect function
